+++ examples/x.py
@@ -34,7 +34,7 @@ def _match(s):
 
 
 def test_x():
-    assert _match('--a-b').end == 5
+    assert _match('--a-b').end() == 5
     assert _match('1 + 2 + 4 + 5 + 6 + 7 + 8 + 9 + 10 + '
                   '(((((('
                   '11 * 12 * 13 * 14 * 15 + 16 * 17 + 18 * 19 * 20'