+++ test/test__match.py
@@ -17,8 +17,9 @@ OneTwoRule = Rule(Sequence(Capture(Literal('1')), Capture(Literal('2'))),
 def test_Match_atom():
     m = Match('123', 0, 1, One, (), {})
     assert m.string == '123'
-    assert m.pos == 0
-    assert m.end == 1
+    assert m.start() == 0
+    assert m.end() == 1
+    assert m.span() == (0, 1)
     assert m.pe is One
     assert m.group(0) == '1'
     assert m.groups() == ()
@@ -29,8 +30,9 @@ def test_Match_atom():
 def test_Match_capture_atom():
     m = Match('123', 0, 1, CaptureOne, ('1',), {})
     assert m.string == '123'
-    assert m.pos == 0
-    assert m.end == 1
+    assert m.start() == 0
+    assert m.end() == 1
+    assert m.span() == (0, 1)
     assert m.pe is CaptureOne
     assert m.group(0) == '1'
     assert m.group(1) == '1'
@@ -42,8 +44,9 @@ def test_Match_capture_atom():
 def test_Match_iterable():
     m = Match('123', 0, 2, OneTwo, (), {})
     assert m.string == '123'
-    assert m.pos == 0
-    assert m.end == 2
+    assert m.start() == 0
+    assert m.end() == 2
+    assert m.span() == (0, 2)
     assert m.pe is OneTwo
     assert m.group(0) == '12'
     assert m.groups() == ()
@@ -54,8 +57,9 @@ def test_Match_iterable():
 def test_Match_capture_iterable():
     m = Match('123', 0, 2, OneCaptureTwo, ('2',), {})
     assert m.string == '123'
-    assert m.pos == 0
-    assert m.end == 2
+    assert m.start() == 0
+    assert m.end() == 2
+    assert m.span() == (0, 2)
     assert m.pe is OneCaptureTwo
     assert m.group(0) == '12'
     assert m.group(1) == '2'
@@ -67,8 +71,9 @@ def test_Match_capture_iterable():
 def test_Match_iterable_bind():
     m = Match('123', 0, 2, OneBindTwo, (), {'x': None})
     assert m.string == '123'
-    assert m.pos == 0
-    assert m.end == 2
+    assert m.start() == 0
+    assert m.end() == 2
+    assert m.span() == (0, 2)
     assert m.pe is OneBindTwo
     assert m.group(0) == '12'
     assert m.group('x') is None
@@ -80,8 +85,9 @@ def test_Match_iterable_bind():
 def test_Match_iterable_bind_capture():
     m = Match('123', 0, 2, OneBindCaptureTwo, (), {'x': '2'})
     assert m.string == '123'
-    assert m.pos == 0
-    assert m.end == 2
+    assert m.start() == 0
+    assert m.end() == 2
+    assert m.span() == (0, 2)
     assert m.pe is OneBindCaptureTwo
     assert m.group(0) == '12'
     assert m.group('x') == '2'
@@ -93,8 +99,9 @@ def test_Match_iterable_bind_capture():
 def test_Match_iterable_rule():
     m = Match('123', 0, 2, OneTwoRule, (['1', '2'],), {})
     assert m.string == '123'
-    assert m.pos == 0
-    assert m.end == 2
+    assert m.start() == 0
+    assert m.end() == 2
+    assert m.span() == (0, 2)
     assert m.pe is OneTwoRule
     assert m.group(0) == '12'
     assert m.group(1) == ['1', '2']